Moving visualization tools to different file
