+++ sdks/python/apache_beam/runners/portability/fn_api_runner/watermark_manager.py
@@ -197,62 +197,3 @@ class WatermarkManager(object):
   def set_watermark(self, name, watermark):
     element = self._watermarks_by_name[name]
     element.set_watermark(watermark)
-
-  def show(self):
-    try:
-      import graphviz
-    except ImportError:
-      import warnings
-      warnings.warn('Unable to draw pipeline. graphviz library missing.')
-      return
-
-    g = graphviz.Digraph()
-
-    def pcoll_node_name(node):
-      if isinstance(node.name, tuple):
-        return 'PCOLL_%s_%s' % node.name
-      else:
-        return 'PCOLL_%s' % node.name
-
-    def add_node(name, shape=None):
-      if name not in seen_nodes:
-        seen_nodes.add(name)
-        g.node(name, shape=shape)
-
-    def add_links(link_from=None, link_to=None, edge_style="solid"):
-      if link_from and link_to:
-        if (link_to, link_from) not in seen_links:
-          g.edge(link_from, link_to, style=edge_style)
-          seen_links.add((link_to, link_from))
-
-    seen_nodes = set()
-    seen_links = set()
-    for node in self._watermarks_by_name.values():
-      if isinstance(node, WatermarkManager.StageNode):
-        name = 'STAGE_%s...%s' % (node.name[:30], node.name[-30:])
-        add_node(name, 'box')
-      else:
-        assert isinstance(node, WatermarkManager.PCollectionNode)
-        name = pcoll_node_name(node)
-        add_node(name)
-
-    for node in self._watermarks_by_name.values():
-      if isinstance(node, WatermarkManager.StageNode):
-        stage = 'STAGE_%s...%s' % (node.name[:30], node.name[-30:])
-        for pcoll in node.inputs:
-          input_name = pcoll_node_name(pcoll)
-          # Main inputs have a BOLD edge.
-          add_links(link_from=input_name, link_to=stage, edge_style="bold")
-        for pcoll in node.side_inputs:
-          # Side inputs have a dashed edge.
-          input_name = pcoll_node_name(pcoll)
-          add_links(link_from=input_name, link_to=stage, edge_style="dashed")
-      else:
-        assert isinstance(node, WatermarkManager.PCollectionNode)
-        pcoll_name = pcoll_node_name(node)
-        for producer in node.producers:
-          prod_name = 'STAGE_%s...%s' % (
-              producer.name[:30], producer.name[-30:])
-          add_links(link_from=prod_name, link_to=pcoll_name)
-
-    g.render('graph', format='png')
